# [METRICS SDK] Fix hash collision in MetricAttributes