--- a/sdk/include/opentelemetry/sdk/metrics/exemplar/reservoir_cell.h
+++ b/sdk/include/opentelemetry/sdk/metrics/exemplar/reservoir_cell.h
@@ -125,6 +125,7 @@ private:
         res.erase(it);
       }
     }
+    res.UpdateHash();
     return res;
   }