--- a/sdk/include/opentelemetry/sdk/metrics/state/filtered_ordered_attribute_map.h
+++ b/sdk/include/opentelemetry/sdk/metrics/state/filtered_ordered_attribute_map.h
@@ -10,6 +10,7 @@
 #include "opentelemetry/common/key_value_iterable.h"
 #include "opentelemetry/nostd/string_view.h"
 #include "opentelemetry/sdk/common/attribute_utils.h"
+#include "opentelemetry/sdk/common/attributemap_hash.h"
 #include "opentelemetry/version.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
@@ -23,21 +24,46 @@ class FilteredOrderedAttributeMap : public opentelemetry::sdk::common::OrderedAt
 {
 public:
   FilteredOrderedAttributeMap() = default;
+
   FilteredOrderedAttributeMap(
       std::initializer_list<std::pair<nostd::string_view, opentelemetry::common::AttributeValue>>
           attributes)
       : OrderedAttributeMap(attributes)
   {}
+
   FilteredOrderedAttributeMap(const opentelemetry::common::KeyValueIterable &attributes)
       : FilteredOrderedAttributeMap(attributes, nullptr)
   {}
+
   FilteredOrderedAttributeMap(const opentelemetry::common::KeyValueIterable &attributes,
                               const opentelemetry::sdk::metrics::AttributesProcessor *processor);
+
   FilteredOrderedAttributeMap(
       std::initializer_list<std::pair<nostd::string_view, opentelemetry::common::AttributeValue>>
           attributes,
       const opentelemetry::sdk::metrics::AttributesProcessor *processor);
+
+  size_t GetHash() const { return _hash; }
+
+private:
+
+  void UpdateHash()
+  {
+    _hash = GetHashForAttributeMap(*this);
+  }
+
+  size_t _hash = 0UL;
 };
+
+class FilteredOrderedAttributeMapHash
+{
+public:
+  size_t operator()(const opentelemetry::sdk::metrics::FilteredOrderedAttributeMap &attributes) const
+  {
+    return attributes.GetHash();
+  }
+};
+
 }  // namespace metrics
 }  // namespace sdk
 OPENTELEMETRY_END_NAMESPACE