--- a/sdk/include/opentelemetry/sdk/metrics/view/attributes_processor.h
+++ b/sdk/include/opentelemetry/sdk/metrics/view/attributes_processor.h
@@ -87,6 +87,8 @@ public:
           }
           return true;
         });
+
+    result.UpdateHash();
     return result;
   }