Fixes #3060 

## Changes

attributes_hashmap_benchmark shows flat result before and after this change.

For significant contributions please make sure you have completed the following items:

* [x] `CHANGELOG.md` updated for non-trivial changes
* [x] Unit tests have been added
* [ ] Changes in public API reviewed