@ThomsonTan 2025-03-26
> @ThomsonTan - Could you also include the benchmark results from before and after in the comments. A slight regression is acceptable, as it's a tradeoff for fixing collision.

Here are the before/after benchmark result. It looks like that the wall time doubled while the CPU time reduced slightly.

* Before

```bash
./measurements_benchmark
2025-03-26T00:43:08+00:00
Running measurements_benchmark
Run on (16 X 2445.43 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 32768 KiB (x1)
Load Average: 1.85, 3.23, 1.73
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
BM_MeasurementsTest    2044809 ns       148990 ns         3856
```

* After

```bash
measurements_benchmark
2025-03-26T00:44:26+00:00
Running measurements_benchmark
Run on (16 X 2445.43 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 32768 KiB (x1)
Load Average: 1.02, 2.64, 1.65
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
BM_MeasurementsTest    4287068 ns        95297 ns         1000
```