@ThomsonTan 2025-03-26
@lalitb it seems the major perf gap is from the below original code. It doesn't sort the incoming `KeyValueIterator` before hashing and query, and the new code does the sort before the hash lookup.

```cpp
  Aggregation *GetOrSetDefault(const opentelemetry::common::KeyValueIterable &attributes,
                               const AttributesProcessor *attributes_processor,
                               std::function<std::unique_ptr<Aggregation>()> aggregation_callback,
                               size_t hash)
  {
    auto it = hash_map_.find(hash);
    if (it != hash_map_.end())
    {
      return it->second.second.get();
    }
```