@marcalff 2025-03-28
Found no issues on this patch (LGTM), but I don't know the metric aggregation code well enough to approve.
Deferring code review to @lalitb , will merge once approved.